Fixed memory usage and start using my SDLogger

+++ gps.h
@@ -5,15 +5,15 @@
 
 #define BUFSIZE 128
 
-class GPS : public SoftwareSerial {
+class GPS: public SoftwareSerial {
   public:
-    GPS(int rx_pin, int tx_pin);
+    GPS(uint8_t rx_pin, uint8_t tx_pin);
     void start();
     char *get_info();
   private:
     void send_ubx(uint8_t *msg, uint8_t len);
     boolean get_ubx_ack(uint8_t *msg);
-    char *get_pubx_item(int itemnr, char *pubx);
+    char *get_pubx_item(uint8_t itemnr, char *pubx);
 };
 
 #endif
