+++ uno.h
@@ -9,9 +9,9 @@
 class UNO {
   public:
     UNO(void);
-    int get_bufsize();
-    char *get_voltage(int pin, int multiply);
-    char *get_temp(int pin, boolean fahrenheit);
+    uint8_t get_bufsize();
+    char *get_voltage(uint8_t pin, uint8_t multiply);
+    char *get_temp(uint8_t pin, boolean fahrenheit);
     char *float_to_str(float value);
 };
